# api-docs.yaml
openapi: 3.0.1
info:
  title: Instant Messaging API
  description: |
    API REST pour une messagerie instantanée qui comprend la gestion  des utilisateur·rice·s, des conversations et des messages. Cette API est utilisée dans  le cadre des modules 194, 195, 322, et 426 à l'EPAI Fribourg.
  contact:
    name: Jérôme Frossard
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
  description: Generated server url
security:
- BearerAuth: []
...
paths:
  /users:
    get:
      tags:
      - user
      summary: Récupère la liste des utilisateur·rice·s
      description: Récupère la liste des utilisateur·rice·s avec pagination.
      operationId: get-users
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Liste récupérée avec succès
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/UsersResponseDto"
              example:
                _embedded:
                  users:
                  - id: 1
                    username: admin@example.com
                    displayName: Administrator
                    _links:
                      self:
                        href: /api/v1/users/1
                  - id: 11
                    username: sheana@example.com
                    displayName: sheana
                    _links:
                      self:
                        href: /api/v1/users/11
                _links:
                  self:
                    href: /api/v1/users?page=0&size=100
                page:
                  size: 20
                  totalElements: 2
                  totalPages: 1
                  number: 0
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
    post:
      tags:
      - user
      summary: Crée un compte d'utilisateur·rice.
      description: "Crée un compte avec un nom d'utilisateur·rice et un mot de passe.\
        \ Le nom peut être soit une adresse e-mail, soit un numéro de téléphone au\
        \ format E.164."
      operationId: create-users
      requestBody:
        description: Utilisateur·rice à créer
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateDto"
            example:
              username: hari@example.com
              displayName: Hari
              password: epai321
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "201":
          description: Utilisateur·rice créé
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
              example:
                id: 11
                username: sheana@example.com
                displayName: sheana
                _links:
                  self:
                    href: /api/v1/users/11
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
...
  /users/{userId}:
    get:
      tags:
      - user
      summary: Récupère les informations d’un·e utilisateur·rice.
      description: Récupère les informations d'un·e utilisateur·rice à partir de son
        identifiant.
      operationId: get-user
      parameters:
      - name: userId
        in: path
        description: Identifiant de l'utilisateur·rice
        required: true
        schema:
          type: integer
          format: int64
        example: 11
      responses:
        "404":
          description: Utilisateur·rice non trouvé·e
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Utilisateur·rice récupéré avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
              example:
                id: 11
                username: sheana@example.com
                displayName: sheana
                _links:
                  self:
                    href: /api/v1/users/11
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
    delete:
      tags:
      - user
      summary: Supprime un·e utilisateur·rice.
      description: Supprime un·e utilisateur·rice à partir de son identifiant. Cette
        opération est irréversible et entraîne l’anonymisation de tous les messages
        envoyés par cette personne.
      operationId: delete-user
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Utilisateur·rice non trouvé·e
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "204":
          description: Utilisateur·rice supprimé avec succès.
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
    patch:
      tags:
      - user
      summary: Modifie les informations d'un·e utilisateur·rice.
      description: Modifie les informations d'un·e utilisateur·rice à partir de son
        identifiant.
      operationId: update-user
      parameters:
      - name: userId
        in: path
        description: Identifiant de l'utilisateur·rice
        required: true
        schema:
          type: integer
          format: int64
        example: 11
      requestBody:
        description: Utilisateur·rice à créer
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateDto"
            examples:
              Modifier le mot de passe:
                description: Modifier le mot de passe
                value:
                  password: Epai123
              Modifier le nom d'utilisateur·rice:
                description: Modifier le nom d'utilisateur·rice
                value:
                  username: "+41799999123"
              Modifier le nom affichable:
                description: Modifier le nom affichable
                value:
                  displayName: Sheana
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "415":
          description: Unsupported Media Type
        "200":
          description: Données modifiées avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
              example:
                id: 11
                username: sheana@example.com
                displayName: sheana
                _links:
                  self:
                    href: /api/v1/users/11
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
...
components:
  schemas:
    UserCreateDto:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
    Links:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/Link"
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        _links:
          $ref: "#/components/schemas/Links"
...
    UserUpdateDto:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
    UsersResponseDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: "#/components/schemas/UserResponseDto"
        _links:
          $ref: "#/components/schemas/Links"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer